+++ crest_data_systems_netapp_eseries_santricity/CHANGELOG.md
@@ -1,4 +1,10 @@
 # CHANGELOG - crest_data_systems_netapp_eseries_santricity
 
+## 1.1.0 / 2023-04-18
+
+- [Updated] Support and setup information in readme.
+- [Fixed] Minor bug-fixes.
+
 ## 1.0.0 / 2023-01-24
-- [Added] Initial release.
+
+- [Added] NetApp ESeries SANtricity version <=11.73.00.0007 is supported in this release.